Change direction, play with starting everything from code
In what will be a procedurally generated setup, trying to start with the
Unity editor, game objects and all that click-and-drag is tough when
trying to design a structure to build from (great for prototyping).
After experimenting and spending too much time thinking, I'm now setting
up a very minimal scene with a single MonoBehaviour that sets up the
rest of the game.

The other aspect of this is I don't want everything to be hard linked to
Unity's structure (e.g. everything is a MonoBehaviour). I am playing
with testability and those are not testable. So what I'm working with
here is Interfaces and a strict seperation of Unity space and Game
space, while trying to keep available as much of the cool features Unity
offers.